// mobi-client/src/types/types.ts
export type Convenient = {
  id: string;
  name: string;
};
//--------------------------------------//
//RoomDetail
export type RoomDetail = {
  id: string;
  title: string;
  description: string;
  address: {
    id: string;
    street: string;
    ward: {
      id: number;
      name: string;
      district: {
        id: number;
        name: string;
        province: {
          id: number;
          name: string;
        };
      };
    };
  };
  priceMonth: number;
  priceDeposit: number;
  roomLength: number;
  roomWidth: number;
  elecPrice: number;
  waterPrice: number;
  maxPeople: number;
  area: number;
  postStartDate: string;
  postEndDate: string;
  images: Image[];
  typepost: string;
  convenients: Convenient[];
};
//--------------------------------------//
export type Image = {
  id: number;
  url: string;
};
// ...
export type PaginatedResponse<T> = {
  data: T[];
  page: number;
  size: number;
  totalElements: number;
  totalPages: number;
  totalRecords: number;
};
// ...
export type RoomInUser = {
  id: string;
  title: string;
  description: string;
  priceMonth: number;
  area: number;
  postStartDate: string;
  address: {
    id: string;
    street: string;
    ward: {
      id: number;
      name: string;
      district: {
        id: number;
        name: string;
        province: {
          id: number;
          name: string;
        };
      };
    };
  };
  conveniences: Convenient[];
  images: Image[];
  landlord: Landlord;
  postType: string;
  isVip?: boolean;
  mainImage?: string;
  favoriteCount?: number
  viewCount?: number;
  
};
export type LandlordProfile = {
  id: string;
  fullName: string;
  email: string;
  phoneNumber: string;
  avatar: string;
};

export type RequirementRequestRoomDto = {
  idRequirement?: string;
  userId: string;
  roomId: string;
  description: string;
};

export type UpdateRequestRoomDto = {
  id: string;
  description: string;
};

export type Landlord = {
  id: string;
  landlordProfile: LandlordProfile;
};